Fix segmentation fault when closing an element
Reported in issue #19. The error happened when the node passed to elm_close
wasn't an element as it was supposed to.

In the example files reported with the issue, the node causing the crash
was a comment.

Now, elm_close will exit with an error if the node it receives isn't
an element. Some calls that could potentially call it with something else
check the node type before calling it.